requirements.txt
streamlit>=1.36
requests>=2.31
beautifulsoup4>=4.12
openai>=1.40
anthropic>=0.34
google-generativeai>=0.7.0
